backend/requirements.txt
# YOLO Detection Backend Requirements
ultralytics==8.0.196
opencv-python==4.8.1.78
flask==2.3.3
flask-cors==4.0.0
numpy==1.24.3
Pillow==10.0.0
torch==2.0.1
torchvision==0.15.2

